fix parseGithub url utils
