+++ server/tests/unit/upload.unit.test.ts
@@ -28,7 +28,7 @@ let downloadGitHubRepoArchive: any
 import { beforeAll } from 'bun:test'
 
 beforeAll(async () => {
-    const utils = await import('../../src/modules/upload/upload.utils')
+    const utils = await import('../../src/modules/upload/d')
     const downloadzip = await import('../../src/modules/upload/downloadzip')
     parseGitHubRepoUrl = utils.parseGitHubRepoUrl
     downloadGitHubRepoArchive = downloadzip.downloadGitHubRepoArchive
